# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "properties": {
          "code": {
            "description": "Stable, machine-readable error code.",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Per-field problems for validation errors.",
            "title": "Details"
          },
          "message": {
            "description": "Human-readable explanation; may change.",
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Echoed in the X-Request-ID header.",
            "title": "Request Id"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "The envelope every error response uses.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "LoginRequest": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "format": "email",
            "maxLength": 254,
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "PasswordChange": {
        "additionalProperties": false,
        "properties": {
          "current_password": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Current Password",
            "type": "string"
          },
          "new_password": {
            "maxLength": 128,
            "minLength": 12,
            "title": "New Password",
            "type": "string"
          }
        },
        "required": [
          "current_password",
          "new_password"
        ],
        "title": "PasswordChange",
        "type": "object"
      },
      "PasswordResetAccepted": {
        "properties": {
          "message": {
            "description": "The same for every address, so the response never reveals whether an account exists.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "PasswordResetAccepted",
        "type": "object"
      },
      "PasswordResetConfirm": {
        "additionalProperties": false,
        "properties": {
          "password": {
            "maxLength": 128,
            "minLength": 12,
            "title": "Password",
            "type": "string"
          },
          "token": {
            "description": "From the emailed link's #token= fragment.",
            "maxLength": 200,
            "minLength": 20,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "password"
        ],
        "title": "PasswordResetConfirm",
        "type": "object"
      },
      "PasswordResetRequest": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "format": "email",
            "maxLength": 254,
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "PasswordResetRequest",
        "type": "object"
      },
...
      "ProfileUpdate": {
        "additionalProperties": false,
        "properties": {
          "first_name": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "First Name"
          },
          "last_name": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Last Name"
          }
        },
        "title": "ProfileUpdate",
        "type": "object"
      },
...
      "RegisterRequest": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "format": "email",
            "maxLength": 254,
            "title": "Email",
            "type": "string"
          },
          "first_name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "First Name",
            "type": "string"
          },
          "last_name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Last Name",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 12,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "first_name",
          "last_name"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "description": "Access token lifetime in seconds.",
            "title": "Expires In",
            "type": "integer"
          },
          "token_type": {
            "const": "Bearer",
            "default": "Bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "expires_in"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
...
      "UserResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "first_name": {
            "title": "First Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "last_name": {
            "title": "Last Name",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          }
        },
        "required": [
          "id",
          "email",
          "first_name",
          "last_name",
          "role",
          "created_at"
        ],
        "title": "UserResponse",
        "type": "object"
      },
      "UserRole": {
        "enum": [
          "customer",
          "admin"
        ],
        "title": "UserRole",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "PC hardware marketplace with a database-enforced build compatibility engine.",
    "title": "Forge API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "description": "",
        "operationId": "post__api_v1_auth_login",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Exchange credentials for an access token and a refresh cookie.",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "description": "",
        "operationId": "post__api_v1_auth_logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "End this session (the refresh token family in the cookie). Always succeeds.",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/logout-all": {
      "post": {
        "description": "",
        "operationId": "post__api_v1_auth_logout-all",
        "parameters": [],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "End every session for the authenticated user, on every device.",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "description": "",
        "operationId": "get__api_v1_auth_me",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "The authenticated user's profile.",
        "tags": [
          "Auth"
        ]
      },
      "patch": {
        "description": "",
        "operationId": "patch__api_v1_auth_me",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Change the account's name. Email changes are not offered (the address is the login).",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/me/password": {
      "post": {
        "description": "",
        "operationId": "post__api_v1_auth_me_password",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChange"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Change the password, given the current one. Every other session is signed out; this one stays.",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/password-reset": {
      "post": {
        "description": "",
        "operationId": "post__api_v1_auth_password-reset",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordResetAccepted"
                }
              }
            },
            "description": "Accepted"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Email a single-use reset link. The answer is the same whether or not the address has an account (no enumeration), and requests are limited per address and per client.",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/password-reset/confirm": {
      "post": {
        "description": "",
        "operationId": "post__api_v1_auth_password-reset_confirm",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetConfirm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Set a new password with the emailed token. The token works once; every session of the account is signed out, and the account owner is told by email.",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "description": "",
        "operationId": "post__api_v1_auth_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Rotate the refresh cookie and issue a new access token. Reusing a rotated token ends the session.",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "description": "",
        "operationId": "post__api_v1_auth_register",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Created"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Create a customer account.",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "security": [],
...
}